Make corpus module
