+++ src/synthesis/mbrola/mbrola_lib.h
@@ -23,7 +23,7 @@
 
 #pragma once
 
-#include "synthesizer.h"
+#include "../../synthesizer.h"
