+++ src/tihu.h
@@ -79,8 +79,8 @@ typedef bool (*TIHU_PROC_INIT)();
 typedef void (*TIHU_PROC_CLOSE)();
 typedef void (*TIHU_PROC_SET_CALLBACK)(TIHU_CALLBACK, void*);
 typedef void (*TIHU_PROC_LOAD_VOICE)(TIHU_VOICE);
+typedef void (*TIHU_PROC_TAG)(const char*);
 typedef void (*TIHU_PROC_SPEAK)(const char*);
-typedef void (*TIHU_PROC_TAGGER)(const char*);
 typedef void (*TIHU_PROC_STOP)();
 typedef bool (*TIHU_PROC_SET_PARAM)(TIHU_PARAM, int);
 typedef bool (*TIHU_PROC_GET_PARAM)(TIHU_PARAM, int &);
@@ -99,6 +99,7 @@ TIHU_FN_DECLARE bool         tihu_LoadUserDic(const char* userdic_path);
 TIHU_FN_DECLARE bool         tihu_LoadVoice(TIHU_VOICE voice);
 TIHU_FN_DECLARE void         tihu_SetCallback(TIHU_CALLBACK call_back, void* user_data);
 TIHU_FN_DECLARE void         tihu_Speak(const char* text);
+TIHU_FN_DECLARE void         tihu_Tag(const char* text);
 TIHU_FN_DECLARE void         tihu_Stop();
 TIHU_FN_DECLARE bool         tihu_SetParam(TIHU_PARAM param, int value);
 TIHU_FN_DECLARE bool         tihu_GetParam(TIHU_PARAM param, int &value);
