+++ src/tihu_dict/afx_entry.cpp
