+++ src/tihu_dict/afx_manager.cpp
