+++ src/tihu_dict/sfx_entry.cpp
