+++ src/tokenizer/tokenizer.cpp
@@ -20,7 +20,7 @@
 *******************************************************************************/
 #include "char_mapper.h"
 #include "tokenizer.h"
-#include "file_manager.h"
+#include "../file_manager.h"
